refactor: atoms 파일 합치기
