+++ atoms/atomsType.ts
@@ -1,0 +1,9 @@
+export interface IModalState {
+  state: boolean
+  id: string
+}
+
+export interface IConditionState {
+  type: string
+  search: string
+}
