+++ atoms/atomsValue.ts
@@ -1,0 +1,18 @@
+import IForm from '../types/Form'
+
+export const defaultFormValue: IForm = {
+  title: '',
+  content: '',
+  url: '',
+}
+export const key = 'current_cardList'
+
+export const defaultModalState = {
+  id: '',
+  state: false,
+}
+
+export const defaultConditionState = {
+  type: '',
+  search: '',
+}
